+++ StreamChatSwiftUITestsAppTests/Pages/MessageListPage.swift
@@ -26,6 +26,10 @@ class MessageListPage {
     static var list: XCUIElement {
         app.scrollViews.matching(NSPredicate(format: "identifier LIKE 'MessageListView' or identifier LIKE 'MessageListScrollView'")).firstMatch
     }
+    
+    static var listEmpty: XCUIElement {
+        app.otherElements.matching(identifier: "EmptyMessagesView").firstMatch
+    }
 
     static var typingIndicator: XCUIElement {
         app.staticTexts["TypingIndicatorBottomView"].firstMatch